requirements.txt
fastapi>=0.95.0,<0.100.0
uvicorn>=0.22.0,<0.24.0
motor>=3.1.1,<3.4.0
pymongo>=4.3.3,<4.6.0
python-dotenv>=1.0.0,<1.1.0
pydantic>=1.10.8,<1.11.0  # Using pydantic v1 for better compatibility
python-decouple>=3.8,<3.9  # Alternative to pydantic-settings
